+ posponer alarma al día siguiente (cherry picked from commit 843b5788d464ec572d81397fc96a4495a53c063e)

+++ src/reloj.c
@@ -127,7 +127,7 @@ void Reloj_activaAlarma(void)
 }
 
 
-void Reloj_posponAlarma(uint8_t minutos)
+void Reloj_posponAlarma(unsigned minutos)
 {
     self->alarmaPospuesta.tiempo = self->tiempo + minutos*60;
     normalizaSegundos(&self->alarmaPospuesta.tiempo);
